requirements.txt
attrs==20.3.0
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
gunicorn==20.1.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pymongo==3.12.0
pyparsing==2.4.7
pytest==6.2.3
python-dotenv==0.17.0
toml==0.10.2
Werkzeug==1.0.1
